ci: update workflow
