+++ package.json
@@ -54,7 +54,8 @@
     "members:gen": "ts-node scripts/all-members.ts",
     "avatars:gen": "ts-node scripts/avatars.ts",
     "search-meta:gen": "ts-node scripts/get-search-meta.ts",
-    "changelog:gen": "ts-node scripts/changelog.ts",
+    "changelog:gen": "ts-node scripts/changelog-generate.ts",
+    "changelog:write": "ts-node scripts/changelog-write.ts",
     "discord": "ts-node scripts/discord.ts",
     "create:pkg": "ts-node scripts/plopfile.ts",
     "release": "changeset publish",
